pyproject.toml
[tool.poetry]
name = "task-queue-project"
version = "0.1.0"
description = ""
authors = ["Anjali Georance <anjaligeorance@gmail.com>"]
readme = "README.md"
packages = [{ include = "app" }]
 
[tool.poetry.dependencies]
python = ">=3.11,<4.0"
fastapi = ">=0.115.12,<0.116.0"
uvicorn = ">=0.34.2,<0.35.0"
pika = ">=1.3.2,<2.0.0"
pymongo = ">=4.13.0,<5.0.0"
redis = ">=6.1.0,<7.0.0"
python-jose = "^3.3.0"
passlib = { extras = ["bcrypt"], version = "^1.7.4" }
python-multipart = "^0.0.20"

[tool.poetry.group.dev.dependencies]
black = "^25.1.0"
isort = "^6.0.1"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
